move user info handler code from SeekerApplication into own class
