+++ Seeker/SeekerApplication.cs
@@ -269,7 +269,7 @@ namespace Seeker
                         maximumConcurrentSearches: 5,
                         serverConnectionOptions: ServerConnectionOptionsWithKeepAlive,
                         addressResolver: ResolveAddressAsync,
-                        userInfoResolver: UserInfoResponseHandler));
+                        userInfoResolver: UserInfoResponder.HandleRequest));
             #endif
             SetDiagnosticState(PreferencesState.LogDiagnostics);
             SeekerState.SoulseekClient.UserStatisticsChanged += SoulseekClient_UserDataReceived;
@@ -962,54 +962,6 @@ namespace Seeker
 
                 ProgressUpdated?.Invoke(null, new ProgressUpdatedUIEventArgs(relevantItem, wasFailed, fullRefresh, percentComplete, e.Transfer.AverageSpeed));
             }
-        }
-
-        public static Task<UserInfo> UserInfoResponseHandler(string uname, IPEndPoint ipEndPoint)
-        {
-            if (UserListService.Instance.IsUserInIgnoreList(uname))
-            {
-                return Task.FromResult(new UserInfo(string.Empty, 0, 0, false));
-            }
-            string bio = PreferencesState.UserInfoBio ?? string.Empty;
-            byte[] picture = GetUserInfoPicture();
-            int uploadSlots = 1;
-            int queueLength = 0;
-            bool hasFreeSlots = true;
-            if (!PreferencesState.SharingOn) //in my experience even if someone is sharing nothing they say 1 upload slot and yes free slots.. but idk maybe 0 and no makes more sense??
-            {
-                uploadSlots = 0;
-                queueLength = 0;
-                hasFreeSlots = false;
-            }
-
-            return Task.FromResult(new UserInfo(bio,uploadSlots, queueLength, hasFreeSlots, picture));
-        }
-
-        private static byte[] GetUserInfoPicture()
-        {
-            if (PreferencesState.UserInfoPictureName == null || PreferencesState.UserInfoPictureName == string.Empty)
-            {
-                return null;
-            }
-            Java.IO.File userInfoPicDir = new Java.IO.File(ApplicationContext.FilesDir, EditUserInfoActivity.USER_INFO_PIC_DIR);
-            if (!userInfoPicDir.Exists())
-            {
-                Logger.Firebase("!userInfoPicDir.Exists()");
-                return null;
-            }
-
-            Java.IO.File userInfoPic = new Java.IO.File(userInfoPicDir, PreferencesState.UserInfoPictureName);
-            if (!userInfoPic.Exists())
-            {
-                //I could imagine a race condition causing this...
-                Logger.Firebase("!userInfoPic.Exists()");
-                return null;
-            }
-            DocumentFile documentFile = DocumentFile.FromFile(userInfoPic);
-            System.IO.Stream imageStream = ApplicationContext.ContentResolver.OpenInputStream(documentFile.Uri);
-            byte[] picFile = new byte[imageStream.Length];
-            imageStream.Read(picFile, 0, (int)imageStream.Length);
-            return picFile;
         }
 
         private void SoulseekClient_PrivilegedUserListReceived(object sender, IReadOnlyCollection<string> privilegedUsers)
