+++ Seeker/Users/ViewUserInfoActivity.cs
@@ -207,7 +207,7 @@ namespace Seeker
                 // viewing self
                 //for UserData we only care about Online Status, upload speed, file count, and dir count
                 userData = new Soulseek.UserData(UserToView, Soulseek.UserPresence.Online, PreferencesState.UploadSpeed, 0, SeekerState.SharedFileCache?.FileCount ?? 0, SeekerState.SharedFileCache?.DirectoryCount ?? 0, "");
-                userInfo = SeekerApplication.UserInfoResponseHandler(UserToView, null).Result; //the task is already completed.  (task.fromresult).
+                userInfo = UserInfoResponder.HandleRequest(UserToView, null).Result; //the task is already completed.  (task.fromresult).
                 hasPicture = userInfo != null && userInfo.HasPicture && userInfo.Picture != null && userInfo.Picture.Length > 0;
             }
             else if (UserToView != null && RequestedUserInfoHelper.GetInfoForUser(UserToView) != null)
